+++ packages/admin-frontend/components/ImageCropper/Crop.js
@@ -3,8 +3,8 @@
 import React, { Component } from 'react';
 import Cropper from 'react-cropper';
 import 'cropperjs/dist/cropper.css';
-import { parseQuery } from '../lib/parseQuery';
-import type { ImageParameters } from '../types';
+import { parseQuery } from '../../lib/parseQuery';
+import type { ImageParameters } from '../../types';
 
 type Props = {
   imageUrl: string,
