+++ packages/admin-frontend/lib/__tests__/utils.test.js
@@ -1,6 +1,6 @@
 // @flow
 
-import { inputStringEmpty } from '../inputStringEmpty';
+import isEmptyString from '../isEmptyString';
 import { parseQuery } from '../parseQuery';
 
 test('Parsing of valid query with number parameters', () => {
@@ -40,33 +40,33 @@ test('Parsing of invalid query parameters', () => {
 });
 
 test('That " " returns empty string', () => {
-  expect(inputStringEmpty(' ')).toEqual(true);
+  expect(isEmptyString(' ')).toEqual(true);
 });
 
 test('That many spaces returns empty string', () => {
-  expect(inputStringEmpty('              ')).toEqual(true);
+  expect(isEmptyString('              ')).toEqual(true);
 });
 
 test('That tabs returns empty string', () => {
-  expect(inputStringEmpty('            ')).toEqual(true);
+  expect(isEmptyString('            ')).toEqual(true);
 });
 
 test('That empty string returns empty string', () => {
-  expect(inputStringEmpty('')).toEqual(true);
+  expect(isEmptyString('')).toEqual(true);
 });
 
 test('That not empty string returns false', () => {
-  expect(inputStringEmpty('not_empty_string')).toEqual(false);
+  expect(isEmptyString('not_empty_string')).toEqual(false);
 });
 
 test('That string with spaces and tabs does not return empty string', () => {
-  expect(inputStringEmpty('This is a   test    !')).toEqual(false);
+  expect(isEmptyString('This is a   test    !')).toEqual(false);
 });
 
 test('That undefined returns empty string', () => {
-  expect(inputStringEmpty(undefined)).toEqual(true);
+  expect(isEmptyString(undefined)).toEqual(true);
 });
 
 test('That null returns empty string', () => {
-  expect(inputStringEmpty(null)).toEqual(true);
+  expect(isEmptyString(null)).toEqual(true);
 });
