--- packages/admin-frontend/lib/inputStringEmpty.js
@@ -1,9 +1,0 @@
-// @flow
-
-export function inputStringEmpty(inputString: ?string) {
-  return (
-    inputString === undefined ||
-    inputString === null ||
-    inputString.trim() === ''
-  );
-}
