+++ packages/admin-frontend/lib/isEmptyString.js
@@ -1,0 +1,5 @@
+// @flow
+
+export default function isEmptyString(str: ?string) {
+  return str == null || str.trim() === '';
+}
